+++ i18n/en.pot
@@ -5,11 +5,50 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-01-17T09:11:56.041Z\n"
-"PO-Revision-Date: 2025-01-17T09:11:56.042Z\n"
+"POT-Creation-Date: 2025-01-20T10:51:19.841Z\n"
+"PO-Revision-Date: 2025-01-20T10:51:19.842Z\n"
 
 msgid "Untitled dashboard"
 msgstr "Untitled dashboard"
+
+msgid "New dashboard: configure external source (superset)"
+msgstr "New dashboard: configure external source (superset)"
+
+msgid "Could not create dashboard"
+msgstr "Could not create dashboard"
+
+msgid "An unknown error occurred"
+msgstr "An unknown error occurred"
+
+msgid "Save dashboard"
+msgstr "Save dashboard"
+
+msgid "Back"
+msgstr "Back"
+
+msgid "Title"
+msgstr "Title"
+
+msgid "Code"
+msgstr "Code"
+
+msgid "Description"
+msgstr "Description"
+
+msgid "Superset Embed ID"
+msgstr "Superset Embed ID"
+
+msgid "Invalid UUID"
+msgstr "Invalid UUID"
+
+msgid "Options"
+msgstr "Options"
+
+msgid "Show chart controls on dashboard items"
+msgstr "Show chart controls on dashboard items"
+
+msgid "Show filters"
+msgstr "Show filters"
 
 msgid "New dashboard: choose type"
 msgstr "New dashboard: choose type"
@@ -34,21 +73,6 @@ msgstr "Configure source"
 
 msgid "Cancel"
 msgstr "Cancel"
-
-msgid "New dashboard: configure external source (superset)"
-msgstr "New dashboard: configure external source (superset)"
-
-msgid "Could not create dashboard"
-msgstr "Could not create dashboard"
-
-msgid "An unknown error occurred"
-msgstr "An unknown error occurred"
-
-msgid "Save dashboard"
-msgstr "Save dashboard"
-
-msgid "Back"
-msgstr "Back"
 
 msgid "The dashboard couldn't be made available offline. Try again."
 msgstr "The dashboard couldn't be made available offline. Try again."
@@ -290,30 +314,6 @@ msgstr "Resources"
 
 msgid "Reports"
 msgstr "Reports"
-
-msgid "Title"
-msgstr "Title"
-
-msgid "Code"
-msgstr "Code"
-
-msgid "Description"
-msgstr "Description"
-
-msgid "Superset Embed ID"
-msgstr "Superset Embed ID"
-
-msgid "Invalid UUID"
-msgstr "Invalid UUID"
-
-msgid "Options"
-msgstr "Options"
-
-msgid "Show chart controls on dashboard items"
-msgstr "Show chart controls on dashboard items"
-
-msgid "Show filters"
-msgstr "Show filters"
 
 msgid "Visualizations"
 msgstr "Visualizations"
