+++ src/components/ConfigureSupersetEmbeddedDashboardModal/index.js
@@ -1,0 +1,1 @@
+export { CreateSupersetEmbeddedDashboard } from './CreateSupersetEmbeddedDashboard.js'
