+++ src/components/DashboardsBar/CreateDashboardButton/styles/DashboardTypeRadio.module.css
@@ -65,7 +65,7 @@
 
 .container input {
     appearance: none;
-    width: 0;
-    height: 0;
+    inline-size: 0;
+    block-size: 0;
     opacity: 0;
 }
